+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog
 1.2.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- WEB-4041 : Handle new "carre" scale
 
 
 1.2 (2023-10-25)